Move with_zodb decorator to indico.core.db.util
